Fixed a few typoes, updated css for register form, added download template, added restriction for deleting Admin account

+++ Backend/services/userService.js
@@ -68,8 +68,18 @@ const updateUser = async (userId, firstName, lastName, email) => {
 };
 
 const deleteUser = async (userId) => {
-  const query = `DELETE FROM ${tableNames.USERS_TABLE} WHERE id=$1`;
-  await pool.query(query, [userId]);
+  const queryGet = `SELECT email FROM ${tableNames.USERS_TABLE} WHERE id=$1`;
+  const result = await pool.query(queryGet, [userId]);
+  if (result.rows.length === 0) {
+    throw new Error("User not found");
+  }
+  if (result.rows[0].email === "admin@vtsappstim.edu.rs") {
+    const err = new Error("The user with this email cannot be deleted");
+    err.status = 403;
+    throw err;
+  }
+  const queryDelete = `DELETE FROM ${tableNames.USERS_TABLE} WHERE id=$1`;
+  await pool.query(queryDelete, [userId]);
 };
 
 const changePassword = async (userId, newPassword) => {
